feat: classificando o usuário com base nas suas contribuições

+++ back-end/api-learn-link/src/main/java/co/solvers/apilearnlink/domain/views/ReacoesEmComentariosDoUsuario/repository/QtdReacoesComentariosUsuarioViewRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface QtdReacoesComentariosUsuarioViewRepository extends JpaRepository<QtdReacoesComentariosUsuarioView, Integer> {
-    Optional<QtdReacoesComentariosUsuarioView> findByUsuarioId(Integer id);
+    Optional<QtdReacoesComentariosUsuarioView> findByUsuarioId(Long id);
 }
